Move line info data to line_data.js
