+++ src/App.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {HashRouter as Router, Route, Switch} from "react-router-dom";
 import LandingPage from './pages/LandingPage';
+import LinePage from './pages/LinePage';
 import AboutPage from './pages/AboutPage';
 import PeoplePage from './pages/PeoplePage';
 
@@ -16,6 +17,9 @@ class App extends React.Component {
     return (
       <Router>
         <Switch>
+          <Route path='/lines/1'>
+            <LinePage />
+          </Route>
           <Route path='/about'>
             <AboutPage />
           </Route>
