+++ src/components/MobileMenuLineList.js
@@ -5,7 +5,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import * as UTIL from '../util';
-import * as CONSTANTS from '../constants';
+import * as LINE_DATA from '../data/line_data';
 
 class MobileMenuLineList extends React.Component {
   /**
@@ -36,7 +36,7 @@ class MobileMenuLineList extends React.Component {
     return <div
       id='mobile-menu-line-list'
       className={`mobile ${this.props.landing_page_state}`}>
-      {CONSTANTS.LINE_INFO.map((line_info, index) => this.generateLineElement(line_info, index))}
+      {LINE_DATA.LINE_INFO.map((line_info, index) => this.generateLineElement(line_info, index))}
     </div>;
   }
 }
