+++ src/data/line_data.js
@@ -1,0 +1,135 @@
+/**
+ * Contains data for all the lines.
+ */
+
+/** @brief Line names in order, with their respective designers */
+export const LINE_INFO = [
+  {
+    name: 'Gossamer Idle',
+    designers: [
+      'Elise Chapman',
+      'Jessica Lai'
+    ],
+    description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras cursus iaculis leo, quis mollis ligula ultrices condimentum. Donec sagittis, sem vitae scelerisque porta, eros libero volutpat orci, quis placerat sapien lacus at quam. Quisque maximus vel enim sit amet blandit. Curabitur tempus nisl a ex aliquet, et lobortis metus rhoncus. Vestibulum sit amet lobortis augue, viverra feugiat nisi. Ut in ultrices risus. Morbi placerat diam semper dapibus rhoncus. Nam ut convallis dui. Praesent dictum erat tincidunt sapien molestie, sed rhoncus est posuere. Morbi placerat diam semper dapibus rhoncus. Nam ut convallis dui. Praesent dictum erat tincidunt sapien molestie, sed rhoncus.'
+  },
+  {
+    name: 'Heroine',
+    designers: [
+      'Lena Li',
+      'Linda Lin'
+    ]
+  },
+  {
+    name: 'Nexus',
+    designers: [
+      'Annalisa Pao'
+    ]
+  },
+  {
+    name: 'Chrysophyta',
+    designers: [
+      'Iz Horgan'
+    ]
+  },
+  {
+    name: 'Mirror',
+    designers: [
+      'Khushi Nandgoankar',
+      'Kathy Zhang'
+    ]
+  },
+  {
+    name: 'The Marianas Trench',
+    designers: [
+      'Lily Oberman',
+      'Skylar McAuliffe'
+    ]
+  },
+  {
+    name: 'Obfuscate',
+    designers: [
+      'Arushi Bandi',
+      'Maya Pandagurun'
+    ]
+  },
+  {
+    name: 'Salvos',
+    designers: [
+      'Patricia Yu'
+    ]
+  },
+  {
+    name: 'Bloom',
+    designers: [
+      'Natalie Waldram',
+      'Amal Jafrani'
+    ]
+  },
+  {
+    name: 'Rensaci',
+    designers: [
+      'Aadya Bhartia',
+    ]
+  },
+  {
+    name: 'Bound',
+    designers: [
+      'Owen Haft',
+      'Carmen Yu',
+    ]
+  },
+  {
+    name: 'Invasion',
+    designers: [
+      'Olivia Luk',
+    ]
+  },
+  {
+    name: 'Somatic',
+    designers: [
+      'Jessa Westheimer'
+    ]
+  },
+  {
+    name: 'Relinquo',
+    designers: [
+      'Kwins Yang',
+      'Jacky Tian'
+    ]
+  },
+  {
+    name: 'Digital Warfare',
+    designers: [
+      'Cassie Scheirer',
+      'Madi Davis'
+    ]
+  },
+  {
+    name: 'Icarus',
+    designers: [
+      'Vincent Derienzo',
+      'Leah Kendrick',
+      'Tyng Peck'
+    ]
+  }
+];
+
+
+export const LINE_NAME_LOOKUP = [
+  'Gossamer Idle',
+  'Heroine',
+  'Nexus',
+  'Chrysophyta',
+  'Mirror',
+  'The Marianas Trench',
+  'Obfuscate',
+  'Salvos',
+  'Bloom',
+  'Rensaci',
+  'Bound',
+  'Invasion',
+  'Somatic',
+  'Relinquo',
+  'Digital Warfare',
+  'Icarus'
+];
