+++ src/styles/desktop/pages/line_page.scss
@@ -1,0 +1,3 @@
+/**
+ * Styles for the page showing each individual line
+ */
